#10825 Refactored PaellaRepository and moved logic to service.
