+++ Resources/config/services.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+
+  <services>
+
+    <service id="pumukitpaellaplayer.paelladata" class="Pumukit\PaellaPlayerBundle\Services\PaellaDataService">
+      <!--<argument type="service" id="request_stack" />-->
+      <argument type="service" id="pumukitschema.pic" />
+      <argument type="service" id="pumukit_baseplayer.trackurl" />
+      <argument type="service" id="mobile_detect.mobile_detector" />
+      <argument type="service" id="pumukit_web_tv.useragent_parser" />
+      <call method="setOpencastClient">
+        <argument type="service" id="pumukit_opencast.client" on-invalid="ignore"/>
+      </call>
+    </service>
+
+  </services>
+</container>
